% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_midi_file.R
\name{read_midi_file}
\alias{read_midi_file}
\title{Process MIDI Information}
\usage{
read_midi_file(midi_file_path)
}
\arguments{
\item{midi_file}{Character. Path to the MIDI file.}
}
\value{
A list with the following components:
\describe{
\item{note_on_events}{Data frame or list. Contains information about note-on events,
with each row representing an event (e.g., note, velocity, time).}
\item{note_off_events}{Data frame or list. Contains information about note-off events,
with each row representing an event (e.g., note, time).}
}
}
\description{
This function returns a list containing MIDI event information with \code{note_on}
and \code{note_off} events extracted from a MIDI file.
}
\examples{
midi_info <- process_midi("example.mid")
midi_info$note_on_events   # View note-on events
midi_info$note_off_events  # View note-off events
}
